+++ src/config.ini
@@ -38,9 +38,12 @@ gpu_number=0
 ## height = video's height
 
 full_video=1
+audio_layer=0:1
+
+
 time_frame=00:00:00
 duration=00:00:10
-audio_layer=0:1
+
 
 # Dandere2x Settings #
